pyproject.toml
[project]
name = "agentic-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.6",
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.11",
    "langchain-google-genai>=4.2.0",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.9",
    "langchain-text-splitters>=1.1.0",
    "matplotlib>=3.10.8",
    "pymupdf>=1.26.7",
    "pypdf>=6.7.0",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
]
